moonboard: drop the Problem property from moonTick
The problem ID is not needed to unpack the tick as the ID is used
as the key to the ticks map/object.  Tocks (pending ticks) in the
app can't rely on a constant ID and so must use the name or URL.
So, drop the ID so that ticks and tocks are interchangeable; if it
turns out a tick needs to store info about the problem the URL can
set by the app when it unpacks the data.