moonboard: split index into problems and setters
Seperate problems and setters in the index, and move the setters part
back into the main local storage.  Originally, the thought was that
smushing everything into a single index would be beneficial for search,
but in practice it's preferable to isolate setters via an explicit
search command.

Problems is currently at ~4.7mb, so there isn't a ton of headroom, but
at the same time I'm not planning on updating much more, and the new
problem set should be hitting soon too.