server: fix checksum caching issue
Originally, all local storage entries were updated on-demand.  This
added a lot of complexity as keeping track of when things were ready
became quite painful.  Updates were then moved to happen immediately
after loading, using promises to enforce ordering.  But, when making
that change a remnant of the error handling was left behind: doUpdate.
Remove the call to doUpdate and flush the checksums if we get rejected
by the substorage.  This should really only happen if there is a truly
bizarre network error, or if a bug gets introduced... :)